Update to layout
Still need to get the form working right and dance selection. Blog is
not up and more work in the work section. module Resume needs to be
created. adds need to be installed and lastly the alert for learn more
needs to setup! Go Luck with the school year, Josh! How fully you end
up in California and Hopefully you can start a great company!